+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.4.16] - 2022-10-29
+- Added support for Python v3.11. Use latest importlib-resources API (`files()` etc.) to fix
+warnings. Fix pylint `__repr__()` usage warnings in tests. Use isort to sort imports and check in
+code formatting action (#85).
+
 ## [0.4.15] - 2022-09-08
 - Fix normal precision inventory unit tests for `decay()` and `cumulative_decays()` methods (#84).
 The tests now warn if the calculated floats are not exact matches for the test defaults, and assert